bm25: refactor the tests, split in files
